Removed 'Json file not found bug'

+++ password.py
@@ -4,19 +4,22 @@ from tkinter import messagebox
 import json
 import string
 import random
+import os
+
+
+#https://stackoverflow.com/questions/61607367/how-to-encrypt-json-in-python
 
 
 window = Tk() # instantiate of a window
-window.geometry("600x400") #dimensions of the window
-window.title("Password Manager") 
+window.geometry("600x400")
+window.title("Password Manager")
 window.config(background="#F7672D")
 
-#Declaration of some objects
 site_text=StringVar()
 password_text=StringVar()
 uname_text=StringVar()
 
-#function to copy password into your clipboard
+
 def copy_button():
 
     if password_text.get()=="":
@@ -28,7 +31,6 @@ def copy_button():
     clip.clipboard_append(password_text.get())  
     clip.destroy()
 
-#function to generate password of length 15
 def generate():
     characters = list(string.ascii_letters + string.digits + "!@#$%^&*()")
     length = 15
@@ -45,13 +47,13 @@ def generate():
     random.shuffle(password)
     password_text.set("".join(password))
 
-#function to save data into a JSON file
+
+
 def Save():
     name=site_text.get()
     password=password_text.get()
     uname=uname_text.get()
 
-#Error message
     if password=="" or name=="" or uname=="":
         messagebox.showerror("Error","Site or password field cannot be empty")
     
@@ -63,15 +65,26 @@ def Save():
             "Password": password
         }
 
- 
-        with open("pass.json","r+") as jfile:
-            try:
-                data=json.load(jfile)
-                data["Details"].append(password_dict)
-                jfile.seek(0)
-                json.dump(data,jfile,indent=2)
-            except JSONDecodeError:
-                pass
+ #update the file with contents not overwrite it
+        if not os.path.exists('/pass.json'):
+                json_obj = {}
+                json_obj['Details'] = []
+                
+                with open('pass.json','w') as jfile:
+                    json.dump(json_obj, jfile)
+                    
+                with open("pass.json","r+") as jfile:
+                    data=json.load(jfile)
+                    data["Details"].append(password_dict)
+                    jfile.seek(0)
+                    json.dump(data,jfile,indent=2)
+        else:
+            with open("pass.json","r+") as jfile:
+                    data=json.load(jfile)
+                    data["Details"].append(password_dict)
+                    jfile.seek(0)
+                    json.dump(data,jfile,indent=2)
+
 
         site_text.set("")
         password_text.set("")
@@ -80,14 +93,21 @@ def Save():
         #save into json file
 
 
-#Function for Opening a new window to create password
+
 def Create_fun():
-    # Toplevel object which will be treated as a new window
+    # Toplevel object which will
+    # be treated as a new window
     newWindow = Toplevel(window)
+ 
+    # sets the title of the
+    # Toplevel widget
     newWindow.title("Create your stuff")
+ 
+    # sets the geometry of toplevel
     newWindow.geometry("600x400")
     newWindow.config(background="#F7672D")
-
+ 
+    # A Label widget to show in toplevel
     Label(newWindow,
     text ="Create a new password").grid(row=0,column=1,pady=10)
 
@@ -103,7 +123,7 @@ def Create_fun():
     Button(newWindow,text="Generate a password " , command=generate,fg="#00FF00",bg="black").grid(row=4,column=1,pady=10)
     Button(newWindow,text="Save",command=Save,fg="#00FF00",bg="black").grid(row=5,column=1,pady=10)
 
-#Function for Opening a new window to save your pasts password
+
 def Save_fun():
  
     newWindow = Toplevel(window)
@@ -124,7 +144,7 @@ def Save_fun():
 
     Button(newWindow,text="Save",command=Save,fg="#00FF00",bg="black",pady=10).grid(row=4,column=1)
     
-#Function for Opening a new window to show password of a particular website
+
 def Show_fun():
     newWindow=Toplevel(window)
     newWindow.title("Show your stuff")
@@ -133,11 +153,9 @@ def Show_fun():
     Label(newWindow,text="Show passwords").grid(row=0,column=1,pady=10)
 
     Label(newWindow,text="Site name : ").grid(row=1,pady=10)
-    #List box to choose a site
     lbox=Listbox(newWindow)
     lbox.grid(row=1,column=2) 
     
-    #storing the site name in a list
     sites=[]
     with open("pass.json","r+") as jfile:
             try:
@@ -148,13 +166,12 @@ def Show_fun():
             except JSONDecodeError:
                 pass
 
-#inserting the website's name into the ListBox
+
     for i in sites:
         lbox.insert(lbox.size(),i)
 
     lbox.config(height=lbox.size())
 
-#Function to show username and password when a selecting a particular website
     def Show():
         Selected=lbox.get(lbox.curselection())
         with open("pass.json","r+") as jfile:
@@ -165,9 +182,7 @@ def Show_fun():
                             password_text.set(i["Password"])
                             uname_text.set(i["Username"])
                 except JSONDecodeError:
-                    pass
-    
-    #site_entry = Entry(newWindow,textvariable = site_text,width=20,state=DISABLED).grid(row=1,column=2) 
+                    pass 
 
     Label(newWindow,text="Username : ").grid(row=2,pady=10)
     uname_entry=Entry(newWindow,textvariable=uname_text,state=DISABLED).grid(row=2,column=2)
@@ -178,8 +193,7 @@ def Show_fun():
     Button(newWindow,text="Show",command=Show,fg="#00FF00",bg="black",pady=10).grid(row=4,column=1,pady=10)
     Button(newWindow,text="Copy Pasword",command=copy_button,fg="#00FF00",bg="black",pady=10).grid(row=5,column=1,pady=10)
 
-
-#Stuff for Main Window  
+    
 Label(window,text="PASSWORD MANAGER",font=(20)).pack(pady=10)
 
 create =Button(window,text="Create",command=Create_fun,fg="#00FF00",bg="black") #remove parenthis of function after placing it
